Enhance dynamic class creation and implement proper date handling
The previous mechanism of class creation in response to JSON data
was found to be lacking in a number of areas when confronted with
real world data so that has been reworked to be more robust. Testing
support put in place along side that.

At the same time processing to identify date/time attributes has been
put in place to turn these into Python Datetime objects. Again testing
support is in place.

Alongside this handling for rate-limiting and time-outs has been added
to the library.