Stop 'reserved name' attributes being created .
The purpose of this change is to allow an object in the form
layer.version to be created while ensuring that, in this case,
an attribute of a Layer object has not been created due to somethign
passed in from the JSON returned from the server.

This change provides a list of names which may not be used to create
class attributes automatically and so disturb 'structural' class nesting
which arises from some objects being embedded within others.