Implementation of 'non-dymanic' instance creation.
Warning this is 3.x only release. A release to cover both
3.x and 2.x will follow later today.

Up to this point the tree of instances which reflected a JSON
response from the server were (apart from the top level class)
"automagically" created in that the resulting instances had no
explicitly defined Class but instead reflected the nature of the
JSON sent . In this commit that no longer happens and instead
a set of explicitly defined classes are used to create the tree
of instances.

There will be another commit soon after this to remove a bit of
crud which I would like commit in case I need to reference it later.