Iterators now create 'non-magic' objects.
In the case of Layer and Publish objects when a user uses the
corresponding iterator to consume multiple objects returned from
the server the objects supplied by the iterator are now 'real'
objects based upon explicitly defined Classes as opposed to objects
which are 'auto-magically' created based only on the contents of the
JSON.

There are other objects that need the same treatment in the future.

One of the motivations for these changes was to allow the fetching
of draft layers and that too is now in place.