Post "modularization" changes to Sphinx sources
In slicing up api.py into different modules the Sphinx build was broken
these changes fix that.

In addition I have added Exceptions to the documentation and changed the
ordering of documented items so that it's reflects how a person might
wish to read it rather than the ordering within the source.