Loosen requirements. Upgrade `requests`
`responses` (unpinned) in requirements-test was depending on
requests>=2.7, which is fairly old.
However, it also imports from urllib3, which was only unvendored from
requests in 2.16.

This pulls in requests>=2.18 (looks like requests 2.16/2.17 had a few
regressions, so this comfortably avoids them)

Other requirements are loosened because it's a bit obnoxious if you're
importing a library and it starts pinning your deps for you :/